% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mvsampen.R
\name{print.experdyn_mvsampen}
\alias{print.experdyn_mvsampen}
\title{Print method for \code{experdyn_mvsampen} objects}
\usage{
\method{print}{experdyn_mvsampen}(x, ...)
}
\arguments{
\item{x}{An \code{experdyn_mvsampen} object.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for \code{experdyn_mvsampen} objects
}
