% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rqa.R
\name{print.experdyn_rqa}
\alias{print.experdyn_rqa}
\title{Print method for \code{experdyn_rqa} objects}
\usage{
\method{print}{experdyn_rqa}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An \code{experdyn_rqa} object.}

\item{digits}{\verb{[integer(1): 3]}\cr Number of decimal places shown.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a concise summary of RQA results produced by \code{\link[=compute_rqa]{compute_rqa()}}.
}
