% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_statespace.R
\name{print.experdyn_statespace}
\alias{print.experdyn_statespace}
\title{Print method for \code{experdyn_statespace} objects}
\usage{
\method{print}{experdyn_statespace}(x, ...)
}
\arguments{
\item{x}{An \code{experdyn_statespace} object.}

\item{...}{Currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a concise summary of the multivariate embedded state space.
}
